Interactive Annotator: Propagated changes from Interactive Editor
